Fix config errors
