+++ optionals/compat_sog/CfgMagazines.hpp
@@ -1,12 +1,3 @@
 class CfgMagazines {
-    class vn_lmgmag_base;
-    class vn_m60_100_mag: vn_lmgmag_base {
-        ACE_isBelt = 1;
-    };
-    class vn_pk_100_mag: vn_lmgmag_base {
-        ACE_isBelt = 1;
-    };
-    class vn_rpd_100_mag: vn_lmgmag_base {
-        ACE_isBelt = 1;
-    };
+    #include "CfgMagazines\belts.hpp"
 };
