+++ optionals/compat_sog/CfgMagazines/belts.hpp
@@ -1,0 +1,10 @@
+class vn_lmgmag_base;
+class vn_m60_100_mag: vn_lmgmag_base {
+    ACE_isBelt = 1;
+};
+class vn_pk_100_mag: vn_lmgmag_base {
+    ACE_isBelt = 1;
+};
+class vn_rpd_100_mag: vn_lmgmag_base {
+    ACE_isBelt = 1;
+};
