+++ optionals/compat_sog/CfgVehicles.hpp
@@ -12,9 +12,9 @@
 }
 
 class CfgVehicles {
-	#include "CfgVehicles/helicopters.hpp";
-	#include "CfgVehicles/planes.hpp";
-	#include "CfgVehicles/tracked.hpp";
-	#include "CfgVehicles/units.hpp";
-	#include "CfgVehicles/wheeled.hpp";
+	#include "CfgVehicles\helicopters.hpp"
+	#include "CfgVehicles\planes.hpp"
+	#include "CfgVehicles\tracked.hpp"
+	#include "CfgVehicles\units.hpp"
+	#include "CfgVehicles\wheeled.hpp"
 };
