+++ optionals/compat_sog/CfgWeapons.hpp
@@ -1,6 +1,6 @@
 class CfgWeapons {
-	#include "CfgWeapons/helmets.hpp";
-	#include "CfgWeapons/launchers.hpp";
-	#include "CfgWeapons/pistols.hpp";
-	#include "CfgWeapons/uniforms.hpp";
+	#include "CfgWeapons\helmets.hpp"
+	#include "CfgWeapons\launchers.hpp"
+	#include "CfgWeapons\pistols.hpp"
+	#include "CfgWeapons\uniforms.hpp"
 };
